+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     * Improved: Display product addons in email order table.
     * Improved: Optimize tax select in jigoshop settings.
     * Improved: Improve eu vat validation.
+    * Fix: Reports CSV export.
 * 1.17.8 - 2015.09.03:
     * Improved: Shipping country now doesn't change on prage reload.
     * Improved: EU Vat is now avaliable to edit.